refactor: extract fixtures models
